Initial set-up of homepage navbar
Successfully linked all menu options in navbar to corresponding php pages

Achieved a background colour for the navbar following the one in the Hair 4 U website initial design.

I am yet to properly align and space out the logo and menu options in the navbar. More research into the relevant CSS code needs to be done.