Yet to get hamburger icon to toggle side menu
I did have the hamburger menu icon able to successfully toggle the side menu that I had implemented before. I was very happy with how it was looking too and how functional it was! Yet to find a fix. Researching how to get it back working again, have left all necessary styling in the CSS.  The references to the ".menu-left" and ".menu-right" classes in the JavaScript code check out to me as well... On another note, I have the two separate menu classes working well in desktop view, the logo is successfully aligned in between them :smile: