Layout for desktop view navbar finalised
Reverted back to original working side-menu for mobile view as the change to the HTML formatting was causing issues. I have now decided to position the menu options to the right of the navbar in desktop view for simplicity. Will now achieve responsiveness of footer and introduce the carousel for the homepage