More responsiveness
Have changed media query to accommodate for how many menu items can uniformly fit on the navbar. Initially, I had this set to a minimum of 768px for "desktop" version of the site to show. This has now been changed to a minimum of 883px. Have also got the ability now to click anywhere outside of the side menu to close it. Implemented this using JavaScript.