Gallery layout - shopping page
Have successfully implemented a gallery layout for the Hair 4 U website's shopping page. It is responsive but I just need to expand the body's height to keep the colour gradient going down past all product images when the site is in mobile view. Next step after this will be connecting this to a MySQL products database where all stock, their pricing and availability will be kept.