Toggling functionality of hamburger menu is back
Had accidentally overwritten the script.js file and doubled up on script for the slideshow gallery., removing the ability to click on the hamburger menu icon and open up the side menu in mobile view. Recycling my previous working script from Github and writing over script.js with it fixed this issue.