Further tidying up
It was revealed that the database was still corrupt when attempting to restart MySQL in XAMPP following changes to the names of important dependencies such as the website's folder in "htdocs".  To fix this I rebuilt the entire database in phpMyAdmin using screenshots from the old one. All is working fine now :smile: I have also included the updated sql file for the site's database.