Login and registration pages successfully setup
Both are functioning and my PHPMail implementation is working successfully; allowing users to be able to activate their accounts through a confirmation link that is sent to their inbox. Admin page is also in the very early stages of development. Next steps are: adding Google/Facebook sign-in integration for the buttons I have setup for these two options on the login page. These will work with current account registration processes, with the exception there won't be a need to store hashed passwords. Additionally, all users seem to have admin privileges currently so soon I will be following Mr. Dempster's next steps of adding an admin column to the accounts table in the Hair 4 U database to remedy this.