Cart and checkout page session scripts working
The scripts for the cart and checkout pages are almost completed and the use of sessions has given the site the ability to carry over the cart total from the overview page to the checkout form page (with the ability to edit quantities properly now!). I am currently working on implementing a way for session variables to also be used to carry account information and prefill the checkout form's contact details section for end-users who are logged in to make 'checking-out' more convenient for them. After working on this, it will be time to implement the payment gateway (likely using Stripe and POLi's secure portals) as well as an admin page for Melissa to monitor her confirmed orders on. Speaking of the admin page, an admin column also needs to be added to the accounts table in phpMyAdmin.