All products added to database
SQL update required. Next steps will be:
getting cart order total calculations working accurately (with the remove button being functional) and the ability to login/signup from cart page and then be redirected to the checkout form.

After successful implementation of this I will explore API integration for payment and address look-up. However, priority will be taken to implementation of joins with the databases, specifically for the purpose of facilitating orders. In this sense, an order ID will comprise of fetching the customer's ID from that database and the selected products from the products database to know who has ordered what. In practice, Melissa (the client) will be able to use this system to make up a customer's order based on this information before posting it out to them.