Guest checkout working
Order subtotals are also being calculated accurately on both checkout screens. On guest_checkout.php, the information that the end-user enters into the form is being stored in the session, so that when payment is finalised and the order is placed, they will have the option of creating a user profile to directly track the shipping status, etc. of their delivery. I will just need to verify that the end-user won't be allowed to continue to payment without selecting their shipping information from the dropdown as this could cause significant issues...