Admin page being re-worked
The page is being dressed up and is receiving further development on functionality. Soon, it will be able to display graphical information of total revenue made, an overview of the number of site visitors and a quick glance of upcoming appointments for the day and orders that need to be fulfilled. The navbar has been set-up to allow the end-user to edit, delete and add products as well as the ability to manage user account information and see all generated invoices for accounting.