Using joins
Joins are being used to record orders into the MySQL database. This works by linking the accounts table to the products table. A record of all orders needs to be visible in the administrator's dashboard and a table of all orders belonging to an individual customer need to be visible on their end (through the customer's portal). Next steps are verifying that the joins work and getting these tables visible in each respective dashboard/portal. Stripe payment integration will also be coming very soon and successful transactions will initiate an order being recorded into the database.