Further development on customer dashboard
Added column, 'role' to accounts table and authenticate now stores the role of the user - admin or customer in a session. It also defaults to 'customer' for all new accounts that are created and the admin can modify this if a user account needs admin privileges - for whatever reason this might be. Orders page has been adapted to display all orders for admin and customer-specific orders for customers.