Tried to get fancy with appointment booking page
Wanted to get the PHP to reference the appointments page to determine what bookings exist in the database so the user doesn't double book. A hair salon is a hard business to get this to work with as the duration of different services varies. A better solution might be transforming this into a page that sends an appointment request to Melissa where she then confirms a time with the customer via email. This was for the excellence part of showing iterations. I have worked out that I will change the content/layout/visual design of the about page and the hair services page to show usability, accessibility, design considerations etc. It will then be a matter if posting this to Mr Dempster ready for assessment 😄